backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-arango==7.5.8
pydantic>=2.6.0
aiofiles==23.2.1 